Fix compiled in configs for working with umbrella
The problem with compiling in configs is that if you have
umbrella projects that are both setting the same config,
either one will get their compiled config in and the other
will fail. Since you can't introduce circular deps in your
umbrella application you would have to introduce an other
application that all of the apps that need FunWithFlags
would need to depend on.

Fixes: #14