Customize UPSERT options for MySQL adapter
ecto_sql supports UPSERT behaviour for both the MySQL and Postgres
adapters. Postgres requires the `conflict_target` option be supplied
when performing updates, while MySQL does not support it at all.

To allow supporting both of these backends, we'll use the Repo's adapter
to differentiate what kind of database we're talking to, and use that to
generate the appropriate insert options.