In the store tests, the nutification-publishing tests should run with both Redis and Ecto setups.
This was originally the case when the tests (and logic!) were in the low-level persistence adapters. Then in commit 592f4fe the logic and tests were moved to the store module, and the test implementation was generalized, but I forgot to adjust the test tags and the newly moved tests ended up being tagged as redis-only.