Add a custom query option to Ecto queries
When having a setup similar to what's described in the official Ecto
guides on setting up multi tenancy with foreign keys, the queries
originating from FunWithFlags can fail, as the tenant key is not set.
See https://hexdocs.pm/ecto/multi-tenancy-with-foreign-keys.html.

This commit adds a custom query option (`:fun_with_flags`) to allow an
easy detection of these queries and to skip the tenant key check.
The test repo was changed to check for this option and raise if it is
not set.

ecto_query_opts